finalizar carga consulta por interfaceLocalPago.pagarcarga() si pudo realizar o no el pago

+++ Proyecto/TallerJavaEquipo6/src/main/java/org/tallerjava/ModuloCarga/aplicacion/impl/ServicioCargaImpl.java
@@ -127,7 +127,7 @@ public class ServicioCargaImpl implements ServicioCarga {
         repoCarga.save(carga);
         repoCargador.save(cargador);
 
-        long importeTotalCentavos = Math.round((importeEnergia + importeDemora) * 100);
+        int importeTotalCentavos = Math.round((importeEnergia + importeDemora) * 100);
 
         boolean pagoExitoso = interfaceLocalPago.pagarCarga(
                 carga.getIdCLiente(),
